Update menus
Changes File to Server. Adds option to disconnect from a server which closes all targets for that server.
Adds a Chat menu with just a single option to close the current chat. If it's a channel it sends a PART message, if a chat it just closes. If a server target it does nothing as you need to purposefully disconnect from the server